requirements.txt
Django==1.9.5
Pygments==2.0.2
argparse==1.2.1
dj-database-url==0.4.1
dj-static==0.0.6
django-toolbelt==0.0.1
gunicorn==19.4.5
psycopg2==2.6.1
static3==0.6.1
wsgiref==0.1.2
djangorestframework==3.2.4
